requirements.txt
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
geopy==2.2.0
Pillow==8.4.0
